Results comparison implementd
theres a function at the bottom that just prints out the percent difference between two fulltexts based on my tests. I have not tested much or connected it to rare words yet but its there now. it uses cosine similairty test so afaik all tests are weighted the same...I might want to look at that and change that.