Normalized and weighted cosine similarity
Given two texts strings, the code can spit out the similarity, 0-1 using the function compareResults. This is the final result of the writing style test. The AI test will also give a result 0-1. Getting a final result between the two of those is just averaging, maybe with weighting depending on more testing.