Forgot to mention a couple things....
So, each attribute was normalized to 0-1, but that was tricky because some attributes had infinity as their actual upper bound. I basically chose realistic upper bounds for some attributes which means that if we get text without periods, or without spaces, we'd start getting some weird results because it would think it was one really big word for example. The big word would be longer than the maximum length I chose, so the similarity might even end up over 100% or under 0% in a case like that....weird stuff. Anyway, the weights were chosen based on some very light testing. They can be changed easily, so feel free to change them and test them.